+++ FrontEnd/ewhaclub/package.json
@@ -13,7 +13,7 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-responsive": "^9.0.0-beta.5",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.3",
     "styled-reset": "^4.3.4",
